backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.5
alembic==1.12.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
numpy==1.26.2 
scikit-learn==1.3.2  # useful for more advanced vector operations